wip: move the service creation out of the api and into dagger
