+++ app/src/main/java/com/chesire/malime/injection/modules/ServerModule.kt
@@ -5,10 +5,18 @@ import com.chesire.malime.core.api.LibraryApi
 import com.chesire.malime.core.api.SearchApi
 import com.chesire.malime.core.flags.SupportedService
 import com.chesire.malime.core.repositories.Authorization
+import com.chesire.malime.kitsu.BuildConfig
+import com.chesire.malime.kitsu.KITSU_ENDPOINT
+import com.chesire.malime.kitsu.api.KitsuAuthInterceptor
 import com.chesire.malime.kitsu.api.KitsuAuthorizer
 import com.chesire.malime.kitsu.api.KitsuManager
+import com.chesire.malime.kitsu.api.KitsuService
 import dagger.Module
 import dagger.Provides
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Singleton
 
 @Suppress("unused")
@@ -16,7 +24,7 @@ import javax.inject.Singleton
 internal class ServerModule {
     @Singleton
     @Provides
-    fun provideAuthorization(
+    fun providesAuthorization(
         kitsuAuthorizer: KitsuAuthorizer
         // malAuthorizer: MalAuthorizer
     ): Authorization {
@@ -39,4 +47,27 @@ internal class ServerModule {
     // for now we can just return the KitsuManager, as we don't support anything else yet
     @Provides
     fun providesSearchApi(manager: KitsuManager): SearchApi = manager
+
+    @Singleton
+    @Provides
+    fun providesKitsuService(kitsuAuthorizer: KitsuAuthorizer): KitsuService {
+        val httpClient = OkHttpClient()
+            .newBuilder()
+            .addInterceptor(KitsuAuthInterceptor(kitsuAuthorizer))
+
+        if (BuildConfig.DEBUG) {
+            val interceptor: HttpLoggingInterceptor = HttpLoggingInterceptor().apply {
+                this.level = HttpLoggingInterceptor.Level.BODY
+            }
+
+            httpClient.addInterceptor(interceptor)
+        }
+
+        return Retrofit.Builder()
+            .baseUrl(KITSU_ENDPOINT)
+            .client(httpClient.build())
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+            .create(KitsuService::class.java)
+    }
 }
