+++ malime.kitsu/src/main/java/com/chesire/malime/kitsu/KitsuConstants.kt
@@ -1,3 +1,3 @@
 package com.chesire.malime.kitsu
 
-internal const val KITSU_ENDPOINT = "https://kitsu.io/"
+const val KITSU_ENDPOINT = "https://kitsu.io/"
