+++ malime.kitsu/src/main/java/com/chesire/malime/kitsu/api/KitsuApi.kt
@@ -1,47 +1,19 @@
 package com.chesire.malime.kitsu.api
 
 import com.chesire.malime.core.flags.ItemType
-import com.chesire.malime.kitsu.BuildConfig
-import com.chesire.malime.kitsu.KITSU_ENDPOINT
 import com.chesire.malime.kitsu.models.request.LoginRequest
 import com.chesire.malime.kitsu.models.request.RefreshAuthRequest
 import com.chesire.malime.kitsu.models.response.AddItemResponse
 import com.chesire.malime.kitsu.models.response.LibraryResponse
 import com.chesire.malime.kitsu.models.response.LoginResponse
 import com.chesire.malime.kitsu.models.response.UpdateItemResponse
-import okhttp3.OkHttpClient
 import okhttp3.RequestBody
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Call
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 import javax.inject.Inject
 
-class KitsuApi @Inject constructor(authorizer: KitsuAuthorizer) {
+class KitsuApi @Inject constructor(
     private val kitsuService: KitsuService
-
-    init {
-        val httpClient = OkHttpClient()
-            .newBuilder()
-            .addInterceptor(KitsuAuthInterceptor(authorizer))
-
-        if (BuildConfig.DEBUG) {
-            val interceptor: HttpLoggingInterceptor = HttpLoggingInterceptor().apply {
-                this.level = HttpLoggingInterceptor.Level.BODY
-            }
-
-            httpClient.addInterceptor(interceptor)
-        }
-
-        val retrofit = Retrofit.Builder()
-            .baseUrl(KITSU_ENDPOINT)
-            .client(httpClient.build())
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-
-        kitsuService = retrofit.create(KitsuService::class.java)
-    }
-
+) {
     fun login(username: String, password: String): Call<LoginResponse> {
         return kitsuService.login(LoginRequest(username, password))
     }
