--- malime.kitsu/src/main/java/com/chesire/malime/kitsu/api/KitsuManagerFactory.kt
@@ -1,7 +1,0 @@
-package com.chesire.malime.kitsu.api
-
-class KitsuManagerFactory {
-    fun get(authorizer: KitsuAuthorizer): KitsuManager {
-        return KitsuManager(KitsuApi(authorizer), authorizer)
-    }
-}
